Initial npm install
All of the code for this commit was downloaded and installed by calling
'sudo npm install' from the *app* directory.

Once this call to 'sudo npm install' is made, all of the files required
(as defined in package.js) are downloaded and placed in the
node_modules directory.

This should allow you to run the application from the app directory
with the command 'npm start', and then test that you have a server
running on http://localhost:3000