Converted Generated app to CoffeeScript
I updated package.json with all of the dependencies required for a
Sunshine server.  I also changed package.json to start with 'coffee'
rather than 'node'.

This change to CoffeeScript requires the conversion of 'app/bin/www' to
CoffeeScript.  But, the other JavaScript files will work just fine with
the CoffeeScript interpreter.