Sunshine Server
This is the first major commit that is not genereated code.  This
commit sets up the core infrastructure for express, socket.io and node
to all work together seamlessly.

This is also where we set up the server to automatically build from our
coffeescript and stylus source files.  We custom compile Stylus to
include 'nib'.

I also renamed app.coffee to sunshine.coffee as this configuration is
one that I already use at SatoriSunshine.com.

I'm also going ahead and creating a route for "partials" which are HTML
snippets (written in Jade) that are used as components in Angular.js.

I also created a folder inside the 'views' folder, also called
'partials', which will hold the Jade templates.

IMPORTANT: The beef of this commit is app/sunshine.coffee.  With this
file set up, there is very little left to do on the server.  Most of
the server work to be done will revolve around the socket.io backend
api.

But the majority of the code in these types of MEAN applications are
actually on the Client.  So much of our application code will be using
Angular.js on the client.