Wolop Branch Created (alpha)
This is the branch we will build the alpha version of Wolop on.  I've
updated the version number in 'package.json' to officially start our
alpha build.

All code prior to this on the master branch is generic and can be used
to create any other application not specific to Wolop.